feat: adicionada entidade LotAllocation para validação de quantidade de ações

+++ src/main/java/com/simulador/investimentos/dtos/OrderDTO.java
@@ -1,10 +1,11 @@
 package com.simulador.investimentos.dtos;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import com.simulador.investimentos.entity.OrderType;
 
-public record OrderDTO(Long id, String assetSymbol, Long walletId, OrderType type, Integer quantity, Double priceAtExecution, LocalDateTime tradeTime) {
+public record OrderDTO(Long id, String assetSymbol, Long walletId, OrderType type, Integer quantity, BigDecimal priceAtExecution, LocalDateTime tradeTime) { 
 	
 	}
