+++ src/main/java/com/simulador/investimentos/dtos/WalletDTO.java
@@ -1,7 +1,8 @@
 package com.simulador.investimentos.dtos;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-public record WalletDTO(Long id, Double balance, Long userId, List<Long> orderIds) {
+public record WalletDTO(Long id, BigDecimal balance, Long userId, List<Long> orderIds) { // BIG DECIMAL
 
 }
